+++ src/components/FallbackMode/index.ts
@@ -1,1 +1,0 @@
-export { FallbackMode } from './FallbackMode'